Move file-listing under explorer
