+++ src/smc-webapp/project/explorer/file-listing/copy-button.tsx
@@ -1,14 +1,14 @@
 import * as React from "react";
 
-const { Space, Icon } = require("../../r_misc");
+import { Space, Icon } from "../../../r_misc";
 const { Button } = require("react-bootstrap");
 
-export function PublicButton({ on_click }) {
+export function CopyButton({ on_click }) {
   return (
     <span>
       <Space />
       <Button bsStyle="info" bsSize="xsmall" onClick={on_click}>
-        <Icon name="bullhorn" /> <span className="hidden-xs">Public</span>
+        <Icon name="files-o" /> <span className="hidden-xs">Copy</span>
       </Button>
     </span>
   );
