+++ src/smc-webapp/project/explorer/file-listing/directory-row.tsx
@@ -1,14 +1,13 @@
 import * as React from "react";
 import memoizeOne from "memoize-one";
 
-import { ProjectActions } from "../../project_actions";
+import { ProjectActions } from "../../../project_actions";
 
 import { CopyButton } from "./copy-button";
 import { PublicButton } from "./public-button";
 import { FileCheckbox } from "./file-checkbox";
 import { generate_click_for } from "./utils";
-
-const { COLORS, TimeAgo, Tip, Icon } = require("../../r_misc");
+import { COLORS, TimeAgo, Tip, Icon } from "../../../r_misc";
 
 const { Row, Col } = require("react-bootstrap");
 const misc = require("smc-util/misc");
