+++ src/smc-webapp/project/explorer/file-listing/file-row.tsx
@@ -1,19 +1,19 @@
 import * as React from "react";
 import memoizeOne from "memoize-one";
 
-import { ProjectActions } from "../../project_actions";
-import { analytics_event } from "../../tracker";
+import { ProjectActions } from "../../../project_actions";
+import { analytics_event } from "../../../tracker";
 
 import { CopyButton } from "./copy-button";
 import { PublicButton } from "./public-button";
 import { FileCheckbox } from "./file-checkbox";
 import { generate_click_for } from "./utils";
 
-const { COLORS, TimeAgo, Tip, Icon } = require("../../r_misc");
+import { COLORS, TimeAgo, Tip, Icon } from "../../../r_misc";
 
 const { Button, Row, Col } = require("react-bootstrap");
 const misc = require("smc-util/misc");
-const { project_tasks } = require("../../project_tasks");
+const { project_tasks } = require("../../../project_tasks");
 
 interface Props {
   name: string;
