+++ src/smc-webapp/project/explorer/file-listing/help-alert.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 
-
 const { Alert } = require("react-bootstrap");
 
 const help_alert_error_syle: React.CSSProperties = {
