+++ src/smc-webapp/project/explorer/file-listing/public-button.tsx
@@ -1,14 +1,13 @@
 import * as React from "react";
-
-const { Space, Icon } = require("../../r_misc");
+import { Space, Icon } from "../../../r_misc";
 const { Button } = require("react-bootstrap");
 
-export function CopyButton({ on_click }) {
+export function PublicButton({ on_click }) {
   return (
     <span>
       <Space />
       <Button bsStyle="info" bsSize="xsmall" onClick={on_click}>
-        <Icon name="files-o" /> <span className="hidden-xs">Copy</span>
+        <Icon name="bullhorn" /> <span className="hidden-xs">Public</span>
       </Button>
     </span>
   );
