+++ src/smc-webapp/project/explorer/file-listing/terminal-mode-display.tsx
