+++ src/smc-webapp/project/explorer/project-files-new.tsx
@@ -7,7 +7,7 @@
  */
 import * as React from "react";
 import { Configuration } from "./Explorer";
-import { EXTs as ALL_FILE_BUTTON_TYPES } from "../file-listing/utils";
+import { EXTs as ALL_FILE_BUTTON_TYPES } from "./file-listing/utils";
 import { Icon } from "../../r_misc";
 import { analytics_event } from "../../tracker";
