--- src/smc-webapp/project/file-listing/README.md
@@ -1,3 +1,0 @@
-# Only the component to list files and folders
-
-See project_files.cjsx for the full page.
